Fixed byte/int mismatch
Former-commit-id: a8002f09c329e83b008c94c931db39e0fd3e0114

+++ common-tools/clas-io/src/main/java/org/jlab/io/hipo/EvioHipoEvent.java
@@ -171,7 +171,7 @@ public class EvioHipoEvent {
                 hipoADC.setByte("sector", i, (byte) evioBank.getInt("sector",i));
                 hipoADC.setByte("layer",  i, (byte) evioBank.getInt("layer",i));
                 hipoADC.setShort("component",  i, (short) evioBank.getInt("strip",i));
-                hipoADC.setInt("ADC",  i, (byte) evioBank.getInt("ADC",i));
+                hipoADC.setInt("ADC",  i, (int) evioBank.getInt("ADC",i));
                 hipoADC.setFloat("time",  i, (float) 0);
                 hipoADC.setShort("ped", i, (short) 0);            
             }
@@ -187,7 +187,7 @@ public class EvioHipoEvent {
                 hipoADC.setByte("sector", i, (byte) evioBank.getInt("sector",i));
                 hipoADC.setByte("layer",  i, (byte) evioBank.getInt("layer",i));
                 hipoADC.setShort("component",  i, (short) evioBank.getInt("strip",i));
-                hipoADC.setInt("ADC",  i, (byte) evioBank.getInt("ADC",i));
+                hipoADC.setInt("ADC",  i, (int) evioBank.getInt("ADC",i));
                 hipoADC.setFloat("time",  i, (float) 0);
                 hipoADC.setShort("ped", i, (short) 0);            
             }
